api/requirements.txt
flask~=3.0.1
Flask-SQLAlchemy~=3.0.5
SQLAlchemy~=2.0.29
flask-cors~=4.0.0
flask-restful~=0.3.10
gunicorn~=22.0.0
gevent~=23.9.1
python-dotenv==1.0.0
psycopg2-binary~=2.9.6
pandas==2.2.3
requests==2.32.3
PyMySQL==1.1.1
cryptography==44.0.0
Flask-Migrate==4.0.7
PyMySQL==1.1.1
psycopg2-binary==2.9.10
pyodbc==5.2.0
sqlparse==0.4.4
sqlglot==11.5.5
flask-apispec==0.11.4
apispec==6.3.0
marshmallow==3.19.0
flask-swagger-ui==4.11.1
openpyxl
pyodbc==5.2.0